+++ supvisors/tools/apis/utils.py
@@ -18,40 +18,23 @@
 # ======================================================================
 
 import os
-import re
-
 from argparse import ArgumentParser
 from urllib.parse import urlparse
 
+from supvisors.utils import parse_docstring
 
-# docstring parsing
+
 def get_docstring_description(func) -> str:
     """ Extract the first part of the docstring. """
-    description = []
-    for line in func.__doc__.split('\n'):
-        stripped_line = line.strip()
-        if stripped_line.startswith('@') or stripped_line.startswith(':'):
-            break
-        description.append(stripped_line)
-    return ' '.join(description)
-
-
-SUPERVISOR_PARAM_FORMAT = r'@param\s+(?P<type>[a-z]+)\s+(?P<name>\w+)'
-SUPVISORS_PARAM_FORMAT = r':param\s+(?P<sname>\w+):'
-DOCSTRING_PARAM_PATTERN = re.compile(rf'^(({SUPERVISOR_PARAM_FORMAT})|({SUPVISORS_PARAM_FORMAT}))\s+(?P<desc>.*)$')
+    docstring = parse_docstring(func.__doc__)
+    return docstring[0][4]
 
 
 def get_docstring_parameters(func) -> str:
     """ Extract the parameters from the docstring.
     Supervisor and Supvisors have different formats. """
-    parameters = {}
-    for line in func.__doc__.split('\n'):
-        stripped_line = line.strip()
-        result = DOCSTRING_PARAM_PATTERN.match(stripped_line)
-        if result:
-            param_name = result.group('name') or result.group('sname')
-            parameters[param_name] = result.group('desc')
-    return parameters
+    docstring = parse_docstring(func.__doc__)
+    return {entry[3]: entry[4] for entry in docstring if entry[1] == 'param'}
 
 
 # Argument parsing